Add primative implementation of slicing component
This adds an early version of the classes involved in slicing a 3D model. Currently. the 3D model is transformed into a series of layers, but is not yet ready for printing.

The following is yet to be implemented:
- Adding extra perimeters to the model
- Adding infill
- Making top and bottom layers solid
- Forming a tool path (including travels)
- GCode generation

However, all valid STLs are now sliced into their "actual" represenation (versus their printable representation)