Rough ProtoType
heres a rough prototype. it can currently render the build space and a model. the model can be scaled and translated(currently theirs a bug in translating). the cammera is started in a defalt position and can be rotated around the x and z axis. rendering is achived by passing a PGraphics object of the desired size to be rendered in.